Add more Paz programs
A collection of Paz programs that we intend to use for the "visible" part
of  submission-time testing of you compiler.  Behind the scenes we will use
a large number of other test programs, some of which explore trickier corner
cases of Paz.  Hence you are advised to build a good, large test suite
yourself.